style: track.visible module conforms to pylint, flake8 and balck formatter
